refactor : CarName에 대한 테스트 분리
